+++ src/app/shared/Directives/highlight.directive.spec.ts
@@ -1,11 +1,11 @@
 /* tslint:disable:no-unused-variable */
 
 import { TestBed, async } from '@angular/core/testing';
-import { HighlightDirective } from './highlight.directive';
+import { ChangeBackgroundColorDirective } from './highlight.directive';
 
 describe('Directive: Highlight', () => {
   it('should create an instance', () => {
-    const directive = new HighlightDirective();
+    const directive = new ChangeBackgroundColorDirective();
     expect(directive).toBeTruthy();
   });
 });