+++ changelog.md
@@ -49,3 +49,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add new components - for single shop and for shop list
 - Add actions, reducers and effects for new entity and for router
 - Add selectors for shops and routing
+
+## [0.0.8] - 2021-07-13
+### Added
+- Add new component process-order to create orders
+- Add reactive form to create a new order
+
+## [0.0.9] - 2021-07-15
+### Added
+- Configure unit tests
+- Fix broken tests and remove obsolete tests
+- Add new tests for pipe and product component